+++ kite/src/main/kotlin/com/tang/kite/wrapper/query/QueryWhereWrapper.kt
@@ -5,9 +5,11 @@ import com.tang.kite.wrapper.where.AbstractWhereWrapper
 /**
  * Query where wrapper for [QueryWrapper]
  *
+ * The reason is that Kotlin's List returns List<out T> instead of List<T>, so java.util.List<T> is used as the query result type.
+ *
  * @author Tang
  */
-class QueryWhereWrapper<T>(private val queryWrapper: QueryWrapper<T>) : AbstractWhereWrapper<T, List<T>, QueryWrapper<T>>(queryWrapper, mutableListOf()) {
+class QueryWhereWrapper<T>(private val queryWrapper: QueryWrapper<T>) : AbstractWhereWrapper<T, java.util.List<T>, QueryWrapper<T>>(queryWrapper, mutableListOf()) {
 
     private val joinedClass = mutableListOf<Class<*>>()
 
@@ -28,9 +30,10 @@ class QueryWhereWrapper<T>(private val queryWrapper: QueryWrapper<T>) : Abstract
     /**
      * Execute the query
      */
-    override fun execute(): List<T> {
+    override fun execute(): java.util.List<T> {
         val list = build().baseMapper.queryWrapper(queryWrapper)
-        return list
+        @Suppress("UNCHECKED_CAST")
+        return list as java.util.List<T>
     }
 
     internal fun isMultiTableQuery(): Boolean {